+++ src/main/kotlin/constants/CodeGeneratorConstants.kt
@@ -10,8 +10,6 @@ object CodeGeneratorConstants {
     const val JAVA_SOURCE_FOLDER_NAME = "java"
     const val KOTLIN_SOURCE_FOLDER_NAME = "kotlin"
 
-    const val EMPTY_OBJECT_PROPERTY_NAME = "EMPTY"
-
     const val IMPLEMENTATION_CONFIG_NAME = "implementation"
     const val ANNOTATION_PROCESSOR_CONFIG_NAME = "annotationProcessor"
     const val CLASSPATH_CONFIG_NAME = "classpath"
