+++ src/main/kotlin/utils/extensions/VirtualFileExt.kt
@@ -3,4 +3,4 @@ package utils.extensions
 import com.intellij.json.JsonLanguage
 import com.intellij.openapi.vfs.VirtualFile
 
-fun VirtualFile.isJsonFile(): Boolean = this.extension!!.toUpperCase() == JsonLanguage.INSTANCE.displayName
+fun VirtualFile.isJsonFile(): Boolean = extension!!.toUpperCase() == JsonLanguage.INSTANCE.displayName
